remove nonsense pending tests
These tests were supposed to test the behavior of a contact resource
used directly as a form object in a Rails form. This is not how it should
be used. Instead, a form should have a simple activemodel-conform form
object which handles validation and stuff and then delegates to one or
more CRM resources for loading and saving form data. Most of the
time form objects would contain domain-specific logic, whereas the CRM
resources are merely API clients.

E.g. a contact page form (where a user enters their name, email,
a text message etc. to contact the website owner) would delegate to
Contact.create and Activity.create based on the intended workflow.

E.g. a support case form (where a user opens a support issue on a
product website) would implement a state machine (to guide the several
states of an issue) and delegate to Activity.create.